+++ app/_layout.tsx
@@ -4,18 +4,18 @@ import { SafeAreaProvider } from "react-native-safe-area-context";
 import { CameraContextProvider } from "../context/CameraContext";
 import { FloorplanProvider } from "../context/FloorplanContext";
 import { LoggerProvider } from "../context/LoggerContext";
-import { ToastProvider } from "../context/ToastProvider";
+import { OverlayProvider } from "../context/Overlays";
 
 export default function RootLayout() {
   return (
     <SafeAreaProvider>
       <LoggerProvider>
         <CameraContextProvider>
-          <ToastProvider>
+          <OverlayProvider>
             <FloorplanProvider>
               <Stack screenOptions={{ headerShown: false }} />
             </FloorplanProvider>
-          </ToastProvider>
+          </OverlayProvider>
         </CameraContextProvider>
       </LoggerProvider>
     </SafeAreaProvider>
