+++ components/index/NewMarkerOptionsModal.tsx
@@ -2,7 +2,7 @@ import { Dispatch, SetStateAction } from "react";
 import { Pressable, Text, TouchableOpacity } from "react-native";
 import Animated, { FadeIn, FadeOut } from "react-native-reanimated";
 
-import { Overlay } from "../../context/ToastProvider";
+import { Overlay } from "../../context/Overlays";
 import { styles } from "../../css/indexStyle";
 
 interface NewMarkerOptionsModalProps {
